+++ src/main/java/com/zerobase/moy/response/exception/CustomExceptionAdvise.java
@@ -1,5 +1,6 @@
-package com.zerobase.moy.exception;
+package com.zerobase.moy.response.exception;
 
+import com.zerobase.moy.response.ApiResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -10,9 +11,9 @@ public class CustomExceptionAdvise {
   @ExceptionHandler(CustomException.class)
   public ResponseEntity<?> customExceptionResponseEntity(CustomException e) {
     return ResponseEntity.status(e.getErrorCode().getStatus())
-        .body(CustomException.CustomExceptionResponse.builder()
-            .message(e.getMessage())
-            .code(e.getErrorCode().toString())
-            .build());
+        .body(ApiResponse.builder()
+            .code(e.getErrorCode())
+            .build()
+        );
   }
 }
